type: add yaksa_type_create_[h]vector_x

+++ src/frontend/include/yaksi.h
@@ -99,8 +99,8 @@ typedef struct yaksi_type_s {
 
     union {
         struct {
-            int count;
-            int blocklength;
+            intptr_t count;
+            intptr_t blocklength;
             intptr_t stride;
             struct yaksi_type_s *child;
         } hvector;
@@ -213,8 +213,8 @@ typedef struct {
 
 
 /* function declarations come at the very end */
-int yaksi_type_create_hvector(int count, int blocklength, intptr_t stride, yaksi_type_s * intype,
-                              yaksi_type_s ** outtype);
+int yaksi_type_create_hvector(intptr_t count, intptr_t blocklength, intptr_t stride,
+                              yaksi_type_s * intype, yaksi_type_s ** outtype);
 int yaksi_type_create_contig(intptr_t count, yaksi_type_s * intype, yaksi_type_s ** outtype);
 int yaksi_type_create_dup(yaksi_type_s * intype, yaksi_type_s ** outtype);
 int yaksi_type_create_hindexed(int count, const int *array_of_blocklengths,
